refactor: split package to module, controller, view, util
