+++ README.md
@@ -224,4 +224,4 @@ Earning rate: 2142.8571428571427
 - [x] 사용되지 않는 getMoney() 제거
 - [x] LotteryController 부분을 하나의 메서드로 묶기
 - [x] Winning이 Reward를 반환
-- [ ] 각 클래스에 대한 패키지 분리
+- [x] 각 클래스에 대한 패키지 분리
