+++ src/main/java/lotto/model/Numbers.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class Numbers {
