+++ src/main/java/lotto/util/Const.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 public final class Const {
     public static final int LOTTERY_PRICE = 1000;
