+++ src/test/java/lotto/MoneyTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.model.Money;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
