+++ src/test/java/lotto/ResultViewTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.view.ResultView;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
