+++ src/test/java/lotto/RewardTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.model.Money;
+import lotto.model.Reward;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
